Fix Add-Type Compilation Error
When Puppet is run using the Local System account either via calling
`puppet agent -t` or when invoked by the service, the execution context
is a non-interactive session. When this happens the temp directory is
set to `C:\Windows\system32\config\systemprofile\AppData\Local\Temp`.

When the temp directory is set that way, `Add-Type` cannot compile any
C# code in the `-TypeDefinition` parameter. The reason is problems with
creating the temp files that are required to complete the compilation
process.

It looks like at some point in the past the behavior of
`System.CodeDom.Compiler.TempFileCollection` may have changed. When the
temp directory is incorrectly set, that assembly does not behave
correctly and causes errors.

This fix re-sets the temp directory to a usable location for that
assembly for the duration of the session, and allows `Add-Type` to
compile the required code.